requirements.txt
pandas~=2.3.2
PyYAML~=6.0.3
requests~=2.32.5
geopy~=2.4.1
fastapi~=0.110.0
uvicorn~=0.29.0
click~=8.1.7
matplotlib~=3.8.2
typer~=0.9.0
psycopg2-binary
celery
redis